only test with cabal?
I can't figure out how to tell travis to install a non-default version
of ghc except by using "language: haskell", so that's what I'll use.
Unfortunately this gives me a lot less control; in particular, this
means I will only test with cabal (the regular cabal build, not even the
cabal new-build), not stack.

I would rather test with both stack and cabal, but I can't figure out
how to configure this. Since I will be developping with stack, that side
needs less testing, so if I'm only going to test with a single build
tool on Travis, it should be cabal.